+++ client/src/components/App/Content/InvestigationForm/TabManagement/ExposuresAndFlights/ExposuresAndFlights.tsx
@@ -1,32 +1,19 @@
 import React , { useContext } from 'react';
-import { AddCircle } from '@material-ui/icons';
+import { Divider } from '@material-ui/core';
 import { FormProvider, useForm } from 'react-hook-form';
-import { Collapse, Divider, Typography, IconButton } from '@material-ui/core';
 
-import Toggle from 'commons/Toggle/Toggle';
-import useFormStyles from 'styles/formStyles';
-import FormTitle from 'commons/FormTitle/FormTitle';
-import FieldName from 'commons/FieldName/FieldName';
-import { fieldsNames } from 'commons/Contexts/ExposuresAndFlights';
-import FormRowWithInput from 'commons/FormRowWithInput/FormRowWithInput';
 import { exposureAndFlightsContext } from 'commons/Contexts/ExposuresAndFlights';
 
-import useStyles from './ExposuresAndFlightsStyles';
-import FlightsForm from './FlightsForm/FlightsForm';
-import { useExposuresAndFlights } from './useExposuresAndFlights'
 import PossibleExposure from './Forms/PossibleExposure'; 
 import { EilatOrDeadSea } from './Forms/EilatOrDeadSea';
-
-const addFlightButton: string = 'הוסף טיסה לחול';
+import { BackFromAbroad } from './Forms/BackFromAbroad';
+import { useExposuresAndFlights } from './useExposuresAndFlights';
 
 const ExposuresAndFlights: React.FC<Props> = ({ id }: Props): JSX.Element => {
   const { exposureAndFlightsData, setExposureDataAndFlights } = useContext(exposureAndFlightsContext);
   const { exposures, wereFlights, wereConfirmedExposures, wasInEilat, wasInDeadSea } = exposureAndFlightsData;
 
   const methods = useForm();
-
-  const { fieldContainer } = useFormStyles();
-  const classes = useStyles();
 
   const {
     saveExposure,
@@ -59,55 +46,15 @@ const ExposuresAndFlights: React.FC<Props> = ({ id }: Props): JSX.Element => {
 
           <Divider />
 
-          <div className={classes.subForm}>
-            <FormTitle title='חזרה מחו״ל' />
-
-            <FormRowWithInput testId='wasAbroad' fieldName='האם חזר מחו״ל?'>
-              <Toggle
-                value={wereFlights}
-                onChange={(event, value) => {
-                  if (value !== null) {
-                    onExposuresStatusChange(fieldsNames.wereFlights, value)
-                  }
-                }}
-              />
-            </FormRowWithInput>
-
-            <Collapse
-              in={wereFlights}
-              className={classes.additionalInformationForm}
-            >
-              <div>
-                <FieldName fieldName='פרטי טיסת חזור לארץ:' className={fieldContainer} />
-                {
-                  exposures.map((exposure, index) =>
-                    exposure.wasAbroad &&
-                    <>
-                      <FlightsForm
-                        fieldsNames={fieldsNames}
-                        key={(exposure.id || '') + index.toString()}
-                        exposureAndFlightsData={exposure}
-                        handleChangeExposureDataAndFlightsField={
-                          (fieldName: string, value: any) => handleChangeExposureDataAndFlightsField(index, fieldName, value)
-                        }
-                      />
-                      <Divider />
-                    </>
-                  )
-                }
-                <IconButton
-                  test-id='addFlight'
-                  onClick={() => onExposureAdded(false, true)}
-                  disabled={disableFlightAddition}
-                >
-                  <AddCircle color={disableFlightAddition ? 'disabled' : 'primary'} />
-                </IconButton>
-                <Typography variant='caption'>
-                  {addFlightButton}
-                </Typography>
-              </div>
-            </Collapse>
-          </div>
+          <BackFromAbroad 
+            wereFlights={wereFlights}
+            onExposuresStatusChange={onExposuresStatusChange}
+            exposures={exposures}
+            handleChangeExposureDataAndFlightsField={handleChangeExposureDataAndFlightsField}
+            onExposureAdded={onExposureAdded}
+            disableFlightAddition={disableFlightAddition}
+          />
+          
         </form>
       </FormProvider>
   );
