+++ client/src/components/App/Content/InvestigationForm/TabManagement/ExposuresAndFlights/useFlightsInvestigation.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import server from 'Utils/axios'
-import {Country} from './FlightsForm/FlightFormTypes';
+import {Country} from './Forms/FlightsForm/FlightFormTypes';
 
 interface FlightsInvestigationProps {
     selectedCity: Country | null;
